refactored the utils.go into respective go files for maintainbility
