+++ scripts/test-api.ps1
@@ -15,11 +15,11 @@ Write-Output "|   Testing (pkg/utils)"
 Write-Output "-----------------------------------"
 
 Set-Location utils
-go test -fuzz=Fuzz -fuzztime 10s -v cloud_test.go main.go
-go test -fuzz=Fuzz -fuzztime 10s -v cni_test.go main.go
-go test -fuzz=Fuzz -fuzztime 10s -v name_test.go main.go
-go test -fuzz=Fuzz -fuzztime 10s -v storage_test.go main.go
-go test -fuzz=Fuzz -fuzztime 10s -v distro_test.go main.go
+go test -fuzz=Fuzz -fuzztime 10s -v cloud_test.go fields.go
+go test -fuzz=Fuzz -fuzztime 10s -v cni_test.go fields.go
+go test -fuzz=Fuzz -fuzztime 10s -v name_test.go fields.go
+go test -fuzz=Fuzz -fuzztime 10s -v storage_test.go fields.go
+go test -fuzz=Fuzz -fuzztime 10s -v distro_test.go fields.go
 go test . -v && Set-Location -
 
 Write-Output "-----------------------------------"
