+++ scripts/test-utils.ps1
@@ -15,9 +15,9 @@ Write-Output "|   Testing (pkg/utils)"
 Write-Output "-----------------------------------"
 
 Set-Location utils
-go test -fuzz=Fuzz -fuzztime 10s -v cloud_test.go utils.go
-go test -fuzz=Fuzz -fuzztime 10s -v cni_test.go utils.go
-go test -fuzz=Fuzz -fuzztime 10s -v name_test.go utils.go
-go test -fuzz=Fuzz -fuzztime 10s -v storage_test.go utils.go
-go test -fuzz=Fuzz -fuzztime 10s -v distro_test.go utils.go
+go test -fuzz=Fuzz -fuzztime 10s -v cloud_test.go fields.go
+go test -fuzz=Fuzz -fuzztime 10s -v cni_test.go fields.go
+go test -fuzz=Fuzz -fuzztime 10s -v name_test.go fields.go
+go test -fuzz=Fuzz -fuzztime 10s -v storage_test.go fields.go
+go test -fuzz=Fuzz -fuzztime 10s -v distro_test.go fields.go
 go test . -v && Set-Location -
